+++ src/main/java/command/Command.java
@@ -1,3 +1,9 @@
+package command;
+
+import ui.Ui;
+import storage.Storage;
+import task.TaskList;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
