+++ src/main/java/command/MarkCommand.java
@@ -1,10 +1,16 @@
+package command;
+
+import ui.Ui;
+import storage.Storage;
+import task.Task;
+import task.TaskList;
+
 import java.io.IOException;
 
-// UnmarkCommand.java
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private final int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
@@ -12,8 +18,8 @@ public class UnmarkCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
             Task task = tasks.get(index - 1);
-            task.markAsNotDone();
-            ui.showTaskUnmarked(task);
+            task.markAsDone();
+            ui.showTaskMarked(task);
             storage.save(tasks);
         } catch (IndexOutOfBoundsException | IOException e) {
             ui.showSavingError();
