+++ src/main/java/command/UnmarkCommand.java
@@ -1,9 +1,16 @@
+package command;
+
+import ui.Ui;
+import storage.Storage;
+import task.Task;
+import task.TaskList;
+
 import java.io.IOException;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
     private final int index;
 
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
@@ -11,8 +18,8 @@ public class MarkCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
             Task task = tasks.get(index - 1);
-            task.markAsDone();
-            ui.showTaskMarked(task);
+            task.markAsNotDone();
+            ui.showTaskUnmarked(task);
             storage.save(tasks);
         } catch (IndexOutOfBoundsException | IOException e) {
             ui.showSavingError();
