+++ src/main/java/storage/Storage.java
@@ -1,3 +1,8 @@
+package storage;
+
+import task.Task;
+import task.TaskList;
+
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
