+++ src/main/java/task/TaskType.java
@@ -1,3 +1,5 @@
+package task;
+
 public enum TaskType {
     TODO,
     DEADLINE,
