+++ src/main/java/task/ToDo.java
@@ -1,9 +1,14 @@
+package task;
+
+import task.Task;
+import task.TaskType;
+
 /**
- * Represents a ToDo task.
+ * Represents a task.ToDo task.
  */
 public class ToDo extends Task {
     /**
-     * Constructor for ToDo class.
+     * Constructor for task.ToDo class.
      *
      * @param description Description of the task.
      */
